✨ limit and limit_raise for unique and nunique
This also includes a refactor of what we previously called 'set'
or 'ordered_set'.

This is now renamed to HashMapUnique, which describes what it does:
It maps keys to unique integers between [0, len(map)-1] using a
hashmap.

To support limit with limit_raise=False, we need tasks to stop early
but still return a value.

The default 'limit_raise=True' is consistent with groupby's row_limit
and is much safer behaviour.